Create most of the code to function dropdowning
The goal is to create a function isnide a module I can then make as a
npm package and use in my future projects. So basically, the dropdowing
function creates (when called) event listeners on all elements that
have the dropdown-button class. After that, the button clicked, it will
toggle the hidden class on the next element (the dropdown-content).
I'm still figuring out how to create a new CSS rule from the module,
because you still need to manually add the css rule
 .hidden {display: none}.